Refactor stdout/stderr creation to share it
Refactor the creation of stdout/stderr for redirection to share it with
other code.
